fix: convert to float32 when calculating test metrics
Avoids overflow from bf16 when tabulating metrics

+++ electrolyte_fm/models/lm_finetuning.py
@@ -362,8 +362,8 @@ class LMFinetuning(LightningModule, DeepSpeedMixin):
         )
         masked_metric_update(
             self.test_metrics,
-            preds,
-            batch["target"],
+            preds.to(dtype=torch.float32),
+            batch["target"].to(dtype=torch.float32),
             batch["target_mask"],
             batch["input_ids"],
             batch.get("is_oov", None),
